+++ setup/wizardsteps.class.inc.php
@@ -1787,7 +1787,7 @@ EOF
 								// Check the module selection
 								try {
 									SetupInfo::SetSelectedModules($aModules);
-									$bSelected = ModuleDiscoveryService::GetInstance()->ComputeBooleanExpression($aInfo['auto_select']);
+									$bSelected = ModuleDiscoveryEvaluationService::GetInstance()->EvaluateBooleanExpression($aInfo['auto_select']);
 								}
 								catch (ModuleDiscoveryServiceException $e) {
 									//logged already
@@ -1865,7 +1865,7 @@ EOF
 						try
 						{
 							SetupInfo::SetSelectedModules($aModules);
-							$bSelected = ModuleDiscoveryService::GetInstance()->ComputeBooleanExpression($aModule['auto_select']);
+							$bSelected = ModuleDiscoveryEvaluationService::GetInstance()->EvaluateBooleanExpression($aModule['auto_select']);
 							if ($bSelected)
 							{
 								$aModules[$sModuleId] = true; // store the Id of the selected module
